requirements.txt
Flask==1.1.2
Flask-Admin==1.5.3
Flask-Login==0.5.0
Flask-WTF==0.14.3
mysql-connector==2.2.9
mysql-connector-python==8.0.17
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
yml==0.0.1
